scripts/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2
google-genai==0.2.0
google-generativeai==0.8.3
Pillow==10.2.0
python-dotenv
urllib3
openai

